+++ device-management-application/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [SW-13851] et-powertop: Add support to reset device stats.
 ### Changed
 - [SW-13832] Use DM_CMD_GET_MM_STATS instead of getTraceBufferServiceProcessor
 ### Deprecated